src/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-MySQLdb==2.0.0
Flask-Toastr==0.5.8
flask-validators==1.0
Flask-WTF==1.2.2
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
langid==1.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.7
numpy==2.2.4
packaging==24.2
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
SQLAlchemy==2.0.40
typing_extensions==4.13.0
Werkzeug==3.1.3
WTForms==3.2.1
